added authentication w/ decoded claims & formatted

+++ Services/PostService.cs
@@ -4,20 +4,31 @@
     {
         private readonly IMapper _mapper;
         private readonly DataContext _context;
-        public PostService(IMapper mapper, DataContext context) 
+        private readonly IHttpContextAccessor _http;
+
+        public PostService(IMapper mapper, DataContext context, IHttpContextAccessor http)
         {
             _mapper = mapper;
             _context = context;
+            _http = http;
         }
+
         public async Task<ServiceResponse<GetPostDto>> CreateOne(AddPostDto newPost)
         {
             var serviceResponse = new ServiceResponse<GetPostDto>();
-            var post = _mapper.Map<Post>(newPost);
-             _context.Posts.Add(post);
-            await _context.SaveChangesAsync();
-            serviceResponse.Data = _mapper.Map<GetPostDto>(post);
-            serviceResponse.Success = true;
-            serviceResponse.Message = "Post successfully created.";
+            try {
+                var ownerId = _http.HttpContext.User.FindFirstValue("Id") ?? throw new Exception("User not found.");
+                var post = _mapper.Map<Post>(newPost);
+                post.OwnerId = ownerId;
+                await _context.Posts.AddAsync(post);
+                await _context.SaveChangesAsync();
+                serviceResponse.Data = _mapper.Map<GetPostDto>(post);
+                serviceResponse.Success = true;
+                serviceResponse.Message = "Post successfully created.";
+            } catch (Exception ex) {
+                serviceResponse.Success = false;
+                serviceResponse.Message = ex.Message;
+            }
             return serviceResponse;
         }
 
@@ -26,14 +37,17 @@
             var serviceResponse = new ServiceResponse<string>();
 
             try {
-                var post = await _context.Posts.FindAsync(id) ?? throw new Exception($"Post with Id:{id} not found.");
+                var ownerId = _http.HttpContext.User.FindFirstValue("Id") ?? throw new Exception("User not found.");
+                var post = await _context.Posts.FindAsync(id) ?? throw new Exception("Post not found.");
+
+                if (ownerId != post.OwnerId) { throw new Exception("You are not the owner of the post."); }
+
                 _context.Posts.Remove(post);
                 await _context.SaveChangesAsync();
                 var posts = await _context.Posts.ToListAsync();
                 serviceResponse.Success = true;
                 serviceResponse.Message = "Post successfully deleted.";
-
-            } catch(Exception ex) { 
+            } catch (Exception ex) {
                 serviceResponse.Success = false;
                 serviceResponse.Message = ex.Message;
             }
@@ -48,9 +62,9 @@
                 serviceResponse.Data = posts.Select(c => _mapper.Map<GetPostDto>(c)).ToList();
                 serviceResponse.Success = true;
                 serviceResponse.Message = "Posts received successfully.";
-            }catch(Exception ex) {
+            } catch (Exception ex) {
                 serviceResponse.Success = false;
-                serviceResponse.Message=ex.Message;
+                serviceResponse.Message = ex.Message;
             }
             return serviceResponse;
         }
@@ -59,18 +73,15 @@
         {
             var serviceResponse = new ServiceResponse<GetPostDto>();
             try {
-
                 var post = await _context.Posts.FindAsync(id) ?? throw new Exception($"Post not found.");
                 serviceResponse.Data = _mapper.Map<GetPostDto>(post);
                 serviceResponse.Success = true;
                 serviceResponse.Message = "Post received successfully.";
-
             } catch (Exception ex) {
                 serviceResponse.Success = false;
                 serviceResponse.Message = ex.Message;
             }
             return serviceResponse;
-
         }
 
         public async Task<ServiceResponse<GetPostDto>> UpdateOne(int id, UpdatePostDto updatedPost)
@@ -78,7 +89,10 @@
             var serviceResponse = new ServiceResponse<GetPostDto>();
 
             try {
+                var ownerId = _http.HttpContext.User.FindFirstValue("Id") ?? throw new Exception("User not found.");
                 var postToUpdate = await _context.Posts.FindAsync(id) ?? throw new Exception($"Post not found.");
+
+                if (ownerId != postToUpdate.OwnerId) { throw new Exception("You are not the owner of the post."); }
 
                 foreach (var property in typeof(UpdatePostDto).GetProperties()) {
                     var updatedValue = property.GetValue(updatedPost);
@@ -87,14 +101,13 @@
                         postProperty.SetValue(postToUpdate, updatedValue);
                     }
                 }
-                    _context.Posts.Update(postToUpdate);
-                    await _context.SaveChangesAsync();
+                _context.Posts.Update(postToUpdate);
+                await _context.SaveChangesAsync();
 
-                    serviceResponse.Data = _mapper.Map<GetPostDto>(postToUpdate);
-                    serviceResponse.Success = true;
-                    serviceResponse.Message = "Post successfully updated.";
-                }  catch (Exception ex) {
-
+                serviceResponse.Data = _mapper.Map<GetPostDto>(postToUpdate);
+                serviceResponse.Success = true;
+                serviceResponse.Message = "Post successfully updated.";
+            } catch (Exception ex) {
                 serviceResponse.Success = false;
                 serviceResponse.Message = ex.Message;
             }
